requirements.txt
# Core AI / API
anthropic>=0.25.0
python-dotenv>=1.0.0

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# Presentation / document generation
python-pptx>=0.6.21
python-docx>=1.1.0

# NotebookLM (unofficial community SDK)
notebooklm>=0.1.0
